+++ position.h
@@ -1,13 +1,14 @@
 #ifndef POSITION_H
 #define POSITION_H
+#include <cstddef>
 
 class Position {
-    int line, col;
+    size_t line, col;
 public:
-    Position(int line, int col);
-    int getLine() const;
-    int getCol() const;
-    void setLine(int line);
-    void setCol(int col);
+    Position(size_t line, size_t col);
+    size_t getLine() const;
+    size_t getCol() const;
+    void setLine(size_t line);
+    void setCol(size_t col);
 };
 #endif
